sentiment-api/requirements.txt
Flask
Flask-Script
Flask-PyMongo
gunicorn
python-dotenv
tensorflow==1.14.0
numpy==1.17.4
scipy==1.3.3
scikit-learn==0.21.3
pillow==4.3.0
h5py==2.8.0
keras==2.2.4
datefinder
spacy
emoji
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.0/en_core_web_sm-2.2.0.tar.gz